backend2/pyproject.toml
[project]
name = "backend2"
version = "0.1.0"
description = "FastAPI OCR service for lab and vital trend analytics"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi==0.116.1",
  "uvicorn[standard]==0.35.0",
  "sqlalchemy==2.0.43",
  "psycopg[binary]>=3.2.10,<3.4",
  "alembic==1.16.4",
  "pydantic-settings==2.10.1",
  "python-multipart==0.0.20",
  "pytesseract==0.3.13",
  "Pillow==11.3.0",
  "pypdf==5.9.0",
  "pypdfium2==4.30.0",
  "minio==7.2.11",
  "httpx==0.25.2"
]

[project.optional-dependencies]
dev = [
  "ruff==0.12.11",
  "mypy==1.17.1"
]
triage = [
  "numpy>=1.26,<3",
  "scikit-learn>=1.4,<2",
  "sentence-transformers>=3.0,<4"
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "C4"]

[tool.ruff.format]
quote-style = "double"

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]
